% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGrid.r
\name{mapGrid}
\alias{mapGrid}
\title{XXX}
\usage{
mapGrid(sGDF, sPDF, we = "", ea = "", so = "", no = "",
  gridValName = "fishing", plotTitle = "", numCats = 5,
  paletteCats = "heat.colors", addLegend = TRUE, legendx = "bottomleft",
  legendncol = 1, legendtitle = "fishing activity", plotPoints = FALSE,
  colPoints = 1, legPoints = FALSE, colLand = "sienna",
  addICESgrid = FALSE, addScale = TRUE, outGridFile = "", outPlot = "",
  ...)
}
\description{
XXX
}
