% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NosatAgebyYearLonLat.r
\name{NosatAgebyYearLonLat}
\alias{NosatAgebyYearLonLat}
\title{XXX}
\usage{
NosatAgebyYearLonLat(lfdat = merged.haddock, alk = alk, chrons = chrons,
  what.species = "Melanogrammus aeglefinus", plus.group = 10)
}
\description{
XXX
}
