% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NosatAgebyYearStatSquare.r
\name{NosatAgebyYearStatSquare}
\alias{NosatAgebyYearStatSquare}
\title{XXX}
\usage{
NosatAgebyYearStatSquare(lfdat = merged.haddock, alk = alk,
  chrons = chrons, what.species = "Melanogrammus aeglefinus",
  plus.group = 10)
}
\description{
XXX
}
