% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMapBlobs.r
\name{plotMapBlobs}
\alias{plotMapBlobs}
\title{XXX}
\usage{
plotMapBlobs(input = cpue.dat, what.quarter = 3, what.year = 2010,
  what.cpue = "cpue.by.n", xlim0 = c(-5, 10), ylim0 = c(45, 62),
  scaling.factor = 10)
}
\description{
XXX
}

