% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrawlSurveyGridPlot.r
\name{TrawlSurveyGridPlot}
\alias{TrawlSurveyGridPlot}
\title{XXX}
\usage{
TrawlSurveyGridPlot(dF, nameLon = "Longitude", nameLat = "Latitude",
  nameVarToSum = "", cellsizeX = 0.5, cellsizeY = 0.5, we = "",
  ea = "", so = "", no = "", gridValName = "fishing", plotMap = TRUE,
  plotTitle = "", numCats = 5, paletteCats = "heat.colors",
  addLegend = TRUE, legendx = "bottomleft", legendncol = 1,
  legendtitle = "fishing activity", plotPoints = TRUE, legPoints = FALSE,
  colPoints = 1, colLand = "sienna", addICESgrid = FALSE,
  addScale = TRUE, outGridFile = "", outPlot = "", ...)
}
\description{
XXX
}
